DOC, MAINT: Docstring Updates Another pass through Docstring and made improvements. moved flag classes to their own module. created a _utils module for misc funcs
